WIP: SSH service and rlaunch service woring

+++ imteksimfw/fireworks/utilities/fwrlm_base.py
@@ -1,8 +1,32 @@
 #!/usr/bin/env python3
+#
+# fwrlm_base.py
+#
+# Copyright (C) 2020 IMTEK Simulation
+# Author: Johannes Hoermann, johannes.hoermann@imtek.uni-freiburg.de
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
 """Manages FireWorks rocket launchers and associated scripts as daemons"""
 
 import os
-import signal  # for unix system signal treatment (https://people.cs.pitt.edu/~alanjawi/cs449/code/shell/UnixSignals.htm)
+import signal  # for unix system signal treatment, see
+# https://people.cs.pitt.edu/~alanjawi/cs449/code/shell/UnixSignals.htm
 import sys  # for stdout and stderr
 import daemon  # for detached daemons, tested for v2.2.4
 import datetime  # for generating timestamps
@@ -13,7 +37,7 @@ import psutil  # checking process status
 import socket  # for host name
 import subprocess
 
-from imteksimfw.fwrlm_config import \
+from imteksimfw.fireworks.fwrlm_config import \
   FW_CONFIG_PREFIX, FW_CONFIG_FILE_NAME, FW_AUTH_FILE_NAME, \
   LAUNCHPAD_LOC, LOGDIR_LOC, MACHINE, SCHEDULER, \
   MONGODB_HOST, MONGODB_PORT_REMOTE, MONGODB_PORT_LOCAL, \
@@ -22,13 +46,10 @@ from imteksimfw.fwrlm_config import \
   RECOVER_OFFLINE, RLAUNCH_FWORKER_FILE, QLAUNCH_FWORKER_FILE, QADAPTER_FILE, \
   MULTI_RLAUNCH_NTASKS,  OMP_NUM_THREADS, FWGUI_PORT
 
-
 # define custom error codes
 pid.PID_CHECK_UNREADABLE = "PID_CHECK_UNREADABLE"
 pid.PID_CHECK_ACCESSDENIED = "PID_CHECK_ACCESSDENIED"
 pid.PID_CHECK_RUNNING = "PID_CHECK_RUNNING"
-
-from imteksimfw.fwrlm_config import config_to_dict, config_keys_to_list
 
 class FireWorksRocketLauncherManager():
     """Base class for managing FireWorks-related daemons"""
@@ -106,11 +127,20 @@ class FireWorksRocketLauncherManager():
 
     def shutdown(self, signum, frame):
         """Shut down daemon neatly"""
-        # try to end all child processes
-        self.logger.debug("Recieved signal {}, shutting down..."
-            .format(signum))
-        os.killpg(os.getpgrp(), signal.SIGTERM)
-        sys.exit(0)
+
+        # no signal received yet, create list
+        if not hasattr(self, '_received_signals'):
+            self._received_signals = []
+
+        # don't do anything if this signal is being handled already
+        if signum not in self._received_signals:
+            self._received_signals.append(signum)
+            self.logger.debug("Recieved signal {}, shutting down..."
+                .format(signum))
+
+            # try to end all child processes
+            os.killpg(os.getpgrp(), signal.SIGTERM)
+            sys.exit(0)
 
     def get_pid(self):
         pidfile_path = os.path.join(self.piddir,self.pidfile_name)
@@ -126,7 +156,8 @@ class FireWorksRocketLauncherManager():
 
         return p
 
-    # inspired by https://github.com/mosquito/python-pidfile/blob/d2ac227a5c3635db2541aa8d8e4507bdeb4b34e2/pidfile/pidfile.py#L15
+    # inspired by
+    # https://github.com/mosquito/python-pidfile/blob/master/pidfile/pidfile.py#L15
     def check_daemon(self, raise_exc=True):
         """Check for a running process by PID file
 
@@ -197,6 +228,7 @@ class FireWorksRocketLauncherManager():
 
         self.logger.debug("PID '{:d}' command line '{}'".format(p,' '.join(cmd)))
 
+        # this warning is somewhat obsolete
         if cmd != self.command_line:
             self.logger.debug("PID file process command line")
             self.logger.debug("")
@@ -283,271 +315,3 @@ class FireWorksRocketLauncherManager():
                     pid=p, exc=exc))
 
         return True
-
-# derivatives
-class DummyManager(FireWorksRocketLauncherManager):
-    # generate descriptive pid files:
-
-    @property
-    def pidfile_name(self):
-        return ".dummy.{user:s}@{host:}.pid".format(
-            user=getpass.getuser(), host=socket.gethostname())
-
-    @property
-    def outfile_name(self):
-        return os.path.join(self.logdir_loc,"dummy_{:s}.out"
-            .format(self.timestamp))
-
-    @property
-    def errfile_name(self):
-        return os.path.join(self.logdir_loc,"dummy_{:s}.err"
-            .format(self.timestamp))
-
-    def spawn(self):
-        """Simple system shell dummy while loop for testing purposes"""
-        args = ['while [ True ]; do printf "."; sleep 5; done']
-        self.logger.debug("Evoking '{cmd:s}'".format(cmd=' '.join(args)))
-        p = subprocess.Popen(args,
-            cwd = self.launchpad_loc,
-            shell = True,
-        )
-        outs, errs = p.communicate()
-        self.logger.debug("Subprocess exited with return code = {}"
-             .format(p.returncode))
-
-class SSHAgentManager(FireWorksRocketLauncherManager):
-    @property
-    def pidfile_name(self):
-        return ".ssh_tunnel.{local_port:d}:{remote_user:s}@{remote_host:s}:{remote_port:d}.{user:s}@{host:}.pid".format(
-            user=getpass.getuser(), host=socket.gethostname())
-
-class SSHTunnelManager(FireWorksRocketLauncherManager):
-    @property
-    def pidfile_name(self):
-        return ".ssh_tunnel.{local_port:d}:{remote_user:s}@{remote_host:s}:{remote_port:d}.{user:s}@{host:}.pid".format(
-            user=getpass.getuser(), host=socket.gethostname())
-
-# stubs
-class RLaunchManager(FireWorksRocketLauncherManager):
-    @property
-    def pidfile_name(self):
-        return ".rlaunch.{user:s}@{host:}.pid".format(
-            user=getpass.getuser(), host=socket.gethostname())
-
-    @property
-    def outfile_name(self):
-        return os.path.join(self.logdir_loc,"rlaunch_{:s}.out"
-            .format(self.timestamp))
-
-    @property
-    def errfile_name(self):
-        return os.path.join(self.logdir_loc,"rlaunch_{:s}.err"
-            .format(self.timestamp))
-
-    @property
-    def rlaunch_fworker_file(self):
-        return RLAUNCH_FWORKER_FILE if RLAUNCH_FWORKER_FILE else os.path.join(
-            FW_CONFIG_PREFIX, "{:s}_noqueue_worker.yaml"
-                .format(self.machine.lower()))
-
-    @property
-    def rlaunch_interval(self):
-        return 10  # seconds
-
-
-    def spawn(self):
-        """spawn rlaunch"""
-        args = [ 'rlaunch',
-                 '-l', self.fw_auth_file_path,
-                 '-w', self.rlaunch_worker_file,
-                 '--loglvl', 'DEBUG', 'rapidfire',
-                 '--nlaunches', 'infinite',
-                 '--sleep', self.rlaunch_interval,
-               ]
-        self.logger.info("Evoking '{cmd:s}'".format(cmd=' '.join(args)))
-        p = subprocess.Popen(args, cwd = self.launchpad_loc)
-        outs, errs = p.communicate()
-        self.logger.info("Subprocess exited with return code = {}"
-             .format(p.returncode))
-
-class QLaunchManager(FireWorksRocketLauncherManager):
-    @property
-    def pidfile_name(self):
-        return ".qlaunch.{user:s}@{host:}.pid".format(
-            user=getpass.getuser(), host=socket.gethostname())
-
-    @property
-    def qlaunch_fworker_file(self):
-        return QLAUNCH_FWORKER_FILE if QLAUNCH_FWORKER_FILE else os.path.join(
-            FW_CONFIG_PREFIX, "{:s}_queue_offline_worker.yaml"
-                .format(self.machine.lower()))
-
-    @property
-    def qadapter_file(self):
-        return QADAPTER_FILE if QADAPTER_FILE else os.path.join(
-            FW_CONFIG_PREFIX, "{:s}_{:s}_qadapter_offline.yaml"
-                .format(self.machine.lower(), self.scheduler.lower()))
-
-class RecoverOfflineManager(FireWorksRocketLauncherManager):
-    @property
-    def pidfile_name(self):
-        return ".recover.{user:s}@{host:}.pid".format(
-            user=getpass.getuser(), host=socket.gethostname())
-
-
-daemon_dict = {
-    'dummy': DummyManager,
-    'rlaunch': RLaunchManager,
-}
-
-# CLI commands pendants
-
-def start_daemon(args):
-    fwrlm = daemon_dict[ args.daemon ]()
-    fwrlm.spawn_daemon()
-
-def check_daemon_status(args):
-    """Checks status of daemon and exits.
-
-    Returns:
-        int, sys.exit exit code
-        - 0: daemon running
-        - 3: daemon not running
-        - 4: state unknown
-
-    Exit codes follow `systemctl`'s exit codes, see
-    https://www.freedesktop.org/software/systemd/man/systemctl.html#Exit%20status
-    """
-    logger = logging.getLogger(__name__)
-
-    fwrlm = daemon_dict[ args.daemon ]()
-    stat = fwrlm.check_daemon(raise_exc=False)
-    logger.info("{:s} daemon state: '{}'".format(args.daemon, stat))
-    if stat == pid.PID_CHECK_RUNNING:
-        ret = 0  # success, daemon running
-    elif stat in [pid.PID_CHECK_NOFILE, pid.PID_CHECK_NOTRUNNING]:
-        ret = 3  # failure, daemon not running
-    else:  # pid.PID_CHECK_UNREADABLE or pid.PID_CHECK_ACCESSDENIED
-        ret = 4  # failure, state unknown
-    sys.exit(ret)
-
-def stop_daemon(args):
-    logger = logging.getLogger(__name__)
-
-    fwrlm = daemon_dict[ args.daemon ]()
-    try:
-        stat = fwrlm.stop_daemon()
-    except Exception as exc:  # stopping failed
-        logger.exception(exc)
-        sys.exit(4)
-
-    if stat:  # successfully stopped
-        logger.info("Stopped.")
-    else:  # wasn't running anyway
-        logger.info("No daemon running.")
-    sys.exit(0)
-
-def main():
-    import argparse
-
-    # in order to have both:
-    # * preformatted help text and ...
-    # * automatic display of defaults
-    class ArgumentDefaultsAndRawDescriptionHelpFormatter(
-        argparse.ArgumentDefaultsHelpFormatter, argparse.RawDescriptionHelpFormatter):
-      pass
-
-    parser = argparse.ArgumentParser(description=__doc__,
-        formatter_class = ArgumentDefaultsAndRawDescriptionHelpFormatter)
-
-    # root-level options
-    parser.add_argument('--debug', default=False, required=False,
-                           action='store_true', dest="debug", help='debug flag')
-    parser.add_argument('--verbose', default=False, required=False,
-                           action='store_true', dest="verbose", help='verbose flag')
-
-    parser.add_argument('--log', required=False, nargs='?', dest="log",
-                        default=None, const='fwrlm.log', metavar='LOG',
-                        help='Write log file fwrlm.log, optionally specify log file name')
-
-    # sub-commands
-    subparsers = parser.add_subparsers(help='command', dest='command')
-
-    # start command
-    start_parser = subparsers.add_parser(
-        'start', help='Start daemons.',
-        formatter_class=ArgumentDefaultsAndRawDescriptionHelpFormatter)
-
-    start_parser.add_argument('daemon', type=str,
-          help='Daemon name', metavar='DAEMON',
-          choices=set(daemon_dict.keys()))
-
-    start_parser.set_defaults(func=start_daemon)
-
-    # status command
-    status_parser = subparsers.add_parser(
-        'status', help='Query daemon status.',
-        formatter_class=ArgumentDefaultsAndRawDescriptionHelpFormatter)
-
-    status_parser.add_argument('daemon', type=str,
-          help='Daemon name', metavar='DAEMON',
-          choices=set(daemon_dict.keys()))
-
-    status_parser.set_defaults(func=check_daemon_status)
-
-    # stop command
-    stop_parser = subparsers.add_parser(
-        'stop', help='Stop daemons.',
-        formatter_class=ArgumentDefaultsAndRawDescriptionHelpFormatter)
-
-    stop_parser.add_argument('daemon', type=str,
-          help='Daemon name', metavar='DAEMON',
-          choices=set(daemon_dict.keys()))
-
-    stop_parser.set_defaults(func=stop_daemon)
-
-    args = parser.parse_args()
-
-    # logging
-    logformat  = "%(levelname)s: %(message)s"
-    if args.debug:
-        logformat  = "[%(asctime)s-%(funcName)s()-%(filename)s:%(lineno)s] %(levelname)s: %(message)s"
-        loglevel = logging.DEBUG
-    elif args.verbose:
-        loglevel = logging.INFO
-    else:
-        loglevel = logging.WARNING
-
-    logging.basicConfig(level=loglevel, format=logformat)
-
-    # explicitly modify the root logger (necessary?)
-    logger = logging.getLogger()
-    logger.setLevel(loglevel)
-
-    # remove all handlers
-    for h in logger.handlers: logger.removeHandler(h)
-
-    # create and append custom handles
-    ch = logging.StreamHandler()
-    formatter = logging.Formatter(logformat)
-    ch.setFormatter(formatter)
-    ch.setLevel(loglevel)
-    logger.addHandler(ch)
-
-    if args.log:
-        fh = logging.FileHandler(args.log)
-        fh.setFormatter(formatter)
-        fh.setLevel(loglevel)
-        logger.addHandler(fh)
-
-    if args.command is None:
-        # if no command supplied, print help
-        parser.print_help()
-    elif 'func' not in args:
-        parser.print_help()
-    else:
-        args.func(args)
-
-
-if __name__ == '__main__':
-    main()
